+++ changelog.md
@@ -10,6 +10,8 @@
      Japanese, Polish, Scottish Gaelic, Slovak, Spanish, Ukrainian.
  ### Lua API
    * Allow specifying custom flags (in particular teleport) when using a custom cost function in wesnoth.find_path
+ ### User Interface
+   * Don't show in the sidebar the time of day schedule of a shrouded hex. (issue #3638)
  ### Packaging
    * OpenMP support has been removed. It is no longer an optional build-time dependency.
  ### Units